+++ bin/static/security.js
@@ -5160,4 +5160,4 @@
   };
 })();
 /* ui-bundle: 6d2fb895cc855280ed44cfb83410e0f39a9b9b7aa2547ce4107f07740b77675e */
-/* ui-sources: bb9778699c8b0c0cefa87cd89ffded1408c07cdf93fe4cb3c9f94e232cb8c82a */
+/* ui-sources: 578ff094f9a45b8aad609ab16aca10dda737d7e90812373903d2050260360ef1 */